+++ lib/screens/examples/multiselect_menu_button_signals.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_quiver/screens/common/signal/multiselect.dart';
+import 'package:flutter_quiver/screens/common/signal/autocomplete.dart';
 import 'package:flutter_quiver/screens/common/signal/multiselect_search.dart';
 import 'package:signals/signals_flutter.dart';
+import '../../data/locations.dart';
 
 /// NOTES: Use signals 🤷‍♂️
 /// The easiest implementation!  Uses a StatelessWidget!
@@ -27,6 +29,9 @@ class MultiSelectMenuButtonExample extends StatelessWidget {
   static const route = '/multiselect_menu_button_example';
   static final model =
       SelectionModel(initialSelection: cities.toSignal(), choices: cities);
+  static final model2 = SelectionModel(
+      initialSelection: {'PJM WH RT'}.toSignal(), choices: locations.toSet());
+  static final location = Signal<String?>('');
 
   @override
   Widget build(BuildContext context) {
@@ -40,6 +45,9 @@ class MultiSelectMenuButtonExample extends StatelessWidget {
             padding: const EdgeInsets.all(16.0),
             child: Row(
               children: [
+                ///
+                /// Simple selection (no search)
+                ///
                 SizedBox(
                   width: 400,
                   child: Column(
@@ -68,14 +76,12 @@ class MultiSelectMenuButtonExample extends StatelessWidget {
                     ],
                   ),
                 ),
-                ///
-                ///
-                ///
                 const SizedBox(
                   width: 48,
                 ),
+
                 ///
-                ///
+                /// Multiselect with search (small number of items)
                 ///
                 SizedBox(
                   width: 400,
@@ -83,7 +89,6 @@ class MultiSelectMenuButtonExample extends StatelessWidget {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: <Widget>[
                       const Text('Selection with search'),
-
                       Container(
                           width: 226,
                           decoration: BoxDecoration(
@@ -96,10 +101,6 @@ class MultiSelectMenuButtonExample extends StatelessWidget {
                               width: 226,
                             ),
                           )),
-
-                      ///
-                      ///
-                      ///
                       const SizedBox(
                         height: 500,
                       ),
@@ -107,6 +108,38 @@ class MultiSelectMenuButtonExample extends StatelessWidget {
                           'Currently selected cities: ${model.currentSelection.value.join(', ')}')),
                       Watch((context) => Text(
                           'Selected cities: ${model.selection.value.join(', ')}')),
+                    ],
+                  ),
+                ),
+
+                ///
+                /// Multiselect with search (many items -- performant)
+                ///
+                SizedBox(
+                  width: 500,
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: <Widget>[
+                      const Text('Selection with search'),
+                      Container(
+                        width: 500,
+                        decoration: BoxDecoration(
+                          color: Colors.blueGrey.shade50,
+                          borderRadius: BorderRadius.circular(4.0),
+                        ),
+                        child: Watch(
+                          (_) => AutocompleteUi(
+                            selection: location,
+                            choices: locations.toSet(),
+                            width: 500,
+                          ),
+                        ),
+                      ),
+                      const SizedBox(
+                        height: 500,
+                      ),
+                      Watch((context) => Text(
+                          'Currently selected location: ${location.value}')),
                     ],
                   ),
                 ),